+++ four/src/lambda/id.rs
@@ -1,0 +1,2 @@
+#[derive(Debug, Clone)]
+pub struct LayerVersionPermissionId;
