+++ four/src/lambda/mod.rs
@@ -3,6 +3,8 @@ mod arn;
 mod event_invoke_config;
 mod function;
 mod id;
+mod layer_version;
+mod layer_version_permission;
 mod property;
 pub mod resource;
 
@@ -17,6 +19,15 @@ pub use event_invoke_config::{
 };
 pub use function::{
     FunctionName, FunctionNameError, MemorySize, MemorySizeError, Timeout, TimeoutError,
+};
+pub use id::LayerVersionPermissionId;
+pub use layer_version::{
+    CompatibleArchitectures, CompatibleArchitecturesError, FunctionContent, LayerName,
+    LayerNameError, LayerVersionDescription, LayerVersionDescriptionError, LicenseInfo,
+    LicenseInfoError,
+};
+pub use layer_version_permission::{
+    GetLayerVersionAction, OrganizationId, OrganizationIdError, Principal, PrincipalError,
 };
 pub use property::{
     architecture::{Architecture, Architectures},
