+++ four/src/lambda/resource/layer_version.rs
@@ -1,15 +1,13 @@
-use four_derive::ManagedResource;
-use nutype::nutype;
-use serde::Serialize;
-
 use crate::{
     core::{
-        convert::WillMappable,
         function::{HaveAtt, RefInner, Referenced},
-        service::IAM,
-        Arn, LogicalId,
+        LogicalId,
     },
-    lambda::property::{architecture::Architecture, runtime::Runtime},
+    lambda::{
+        arn::LayerVersionArn, layer_version::CompatibleRuntimes, CompatibleArchitectures,
+        FunctionContent, LayerName, LayerVersionDescription, LicenseInfo,
+    },
+    ManagedResource,
 };
 
 #[derive(ManagedResource, Clone)]
@@ -18,53 +16,11 @@ pub struct LayerVersion {
     logical_id: LogicalId,
     compatible_architectures: Option<CompatibleArchitectures>,
     compatible_runtimes: Option<CompatibleRuntimes>,
-    content: Content,
-    description: Option<Description>,
+    content: FunctionContent,
+    description: Option<LayerVersionDescription>,
     layer_name: Option<LayerName>,
     license_info: Option<LicenseInfo>,
 }
-
-#[nutype(
-    validate(predicate = |architectures| architectures.len() <= 2),
-    derive(Debug, Clone, Serialize),
-)]
-pub struct CompatibleArchitectures(Vec<Architecture>);
-
-#[nutype(
-    validate(predicate = |runtimes| runtimes.len() <= 15),
-    derive(Debug, Clone, Serialize),
-)]
-pub struct CompatibleRuntimes(Vec<Runtime>);
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Content {
-    // TODO: define type in response to Amazon S3 support feature
-    s3_bucket: String,
-    s3_key: String,
-    s3_object_version: String,
-}
-
-#[nutype(
-    validate(len_char_min = 0, len_char_max = 256),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct Description(String);
-
-#[nutype(
-    validate(
-        len_char_min = 1,
-        len_char_max = 140,
-        regex = r#"(arn:[a-zA-Z0-9-]+:lambda:[a-zA-Z0-9-]+:\d{12}:layer:[a-zA-Z0-9-_]+)|[a-zA-Z0-9-_]+"#
-    ),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct LayerName(String);
-
-#[nutype(validate(len_char_max = 512), derive(Debug, Clone, Serialize))]
-pub struct LicenseInfo(String);
-
-#[derive(Debug, Clone, Serialize)]
-pub struct LayerVersionArn(Arn<IAM>);
 
 impl Referenced for LayerVersion {
     type To = LayerVersionArn;
@@ -73,8 +29,6 @@ impl Referenced for LayerVersion {
         RefInner::Id(self.logical_id.clone())
     }
 }
-
-impl WillMappable<LayerName> for LayerVersionArn {}
 
 impl HaveAtt<LayerVersionArn> for LayerVersion {
     const KEY: &'static str = "LayerVersionArn";
